# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/batches/{id}": {
      "get": {
        "operationId": "getBatchById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDto"
                }
              }
            }
          },
          "404": {
            "description": "Batch not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a batch of a published course",
        "tags": [
          "batches"
        ]
      }
    },
    "/api/v1/admin/courses/{courseId}/batches": {
      "get": {
        "operationId": "adminListBatches",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BatchDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Course not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "List the batches of a course",
        "tags": [
          "admin: batches"
        ]
      },
      "post": {
        "operationId": "adminCreateBatch",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; `errors` maps field paths to messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Course not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "Schedule a batch of a course",
        "tags": [
          "admin: batches"
        ]
      }
    },
    "/api/v1/admin/batches/{id}": {
      "get": {
        "operationId": "adminGetBatch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Batch not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "Get a batch",
        "tags": [
          "admin: batches"
        ]
      },
      "patch": {
        "operationId": "adminUpdateBatch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; `errors` maps field paths to messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Batch not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "Update a batch",
        "tags": [
          "admin: batches"
        ]
      },
      "delete": {
        "operationId": "adminDeleteBatch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Batch not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "Delete a batch",
        "tags": [
          "admin: batches"
        ]
      }
    },
...
  },
  "info": {
    "title": "PrincetonBD API",
    "description": "Backend API for the PrincetonBD web application.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
  "components": {
    "securitySchemes": {
      "workos": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "WorkOS AuthKit access token."
      }
    },
    "schemas": {
...
      "BranchDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "dhaka-gulshan"
          },
          "name": {
            "type": "string",
            "example": "Dhaka — Gulshan"
          },
          "address": {
            "type": "string",
            "nullable": true,
            "example": "House 00, Road 00, Gulshan-2, Dhaka 1212"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": "+880 1700-000000"
          },
          "isActive": {
            "type": "boolean",
            "description": "Inactive branches are hidden from public lists but keep their history."
          },
          "sortOrder": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "address",
          "phone",
          "isActive",
          "sortOrder",
          "createdAt",
          "updatedAt"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "Course not found"
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "ValidationErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "Course not found"
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "name": [
                "name should not be empty"
              ],
              "modules.0.title": [
                "title must be a string"
              ]
            }
          }
        },
        "required": [
          "statusCode",
          "message",
          "errors"
        ]
      },
...
      "DeliveryMode": {
        "type": "string",
        "enum": [
          "classroom",
          "live_online"
        ]
      },
      "BatchTeacherDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name"
        ]
      },
      "Weekday": {
        "type": "string",
        "enum": [
          "sat",
          "sun",
          "mon",
          "tue",
          "wed",
          "thu",
          "fri"
        ],
        "description": "Which days of the week the batch meets, Saturday first."
      },
      "BatchStatus": {
        "type": "string",
        "enum": [
          "open",
          "filling",
          "waitlist",
          "closed"
        ]
      },
      "BatchDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "courseId": {
            "type": "string",
            "format": "uuid"
          },
          "courseSlug": {
            "type": "string",
            "example": "sat"
          },
          "courseName": {
            "type": "string",
            "example": "SAT"
          },
          "mode": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DeliveryMode"
              }
            ]
          },
          "branch": {
            "nullable": true,
            "description": "Null for a live-online batch.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/BranchDto"
              }
            ]
          },
          "teacher": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/BatchTeacherDto"
              }
            ]
          },
          "startsOn": {
            "type": "string",
            "description": "Dhaka calendar day, YYYY-MM-DD.",
            "example": "2026-10-12"
          },
          "endsOn": {
            "type": "string",
            "example": "2026-12-18"
          },
          "days": {
            "type": "array",
            "description": "Which days of the week the batch meets, Saturday first.",
            "items": {
              "$ref": "#/components/schemas/Weekday"
            }
          },
          "startTime": {
            "type": "string",
            "description": "Dhaka wall-clock time, HH:MM.",
            "example": "18:30"
          },
          "endTime": {
            "type": "string",
            "example": "20:30"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/BatchStatus"
              }
            ]
          },
          "seatsLeft": {
            "type": "number",
            "nullable": true
          },
          "feeAmount": {
            "type": "number",
            "nullable": true,
            "description": "Whole taka; null means the course price applies."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "courseId",
          "courseSlug",
          "courseName",
          "mode",
          "branch",
          "teacher",
          "startsOn",
          "endsOn",
          "days",
          "startTime",
          "endTime",
          "status",
          "seatsLeft",
          "feeAmount",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateBatchDto": {
        "type": "object",
        "properties": {
          "mode": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DeliveryMode"
              }
            ]
          },
          "branchId": {
            "type": "string",
            "nullable": true,
            "description": "Required for a classroom batch, must be empty for a live-online one.",
            "format": "uuid"
          },
          "teacherId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "startsOn": {
            "type": "string",
            "description": "Dhaka calendar day, YYYY-MM-DD.",
            "example": "2026-10-12"
          },
          "endsOn": {
            "type": "string",
            "example": "2026-12-18"
          },
          "days": {
            "type": "array",
            "uniqueItems": true,
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/Weekday"
            }
          },
          "startTime": {
            "type": "string",
            "description": "Dhaka wall-clock time, HH:MM (24-hour).",
            "example": "18:30"
          },
          "endTime": {
            "type": "string",
            "example": "20:30"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/BatchStatus"
              }
            ]
          },
          "seatsLeft": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "feeAmount": {
            "type": "number",
            "nullable": true,
            "description": "Whole taka; overrides the course price for this run. Null means the course price.",
            "minimum": 0
          }
        },
        "required": [
          "mode",
          "startsOn",
          "endsOn",
          "days",
          "startTime",
          "endTime"
        ]
      },
      "UpdateBatchDto": {
        "type": "object",
        "properties": {
          "mode": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DeliveryMode"
              }
            ]
          },
          "branchId": {
            "type": "string",
            "nullable": true,
            "description": "Required for a classroom batch, must be empty for a live-online one.",
            "format": "uuid"
          },
          "teacherId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "startsOn": {
            "type": "string",
            "description": "Dhaka calendar day, YYYY-MM-DD.",
            "example": "2026-10-12"
          },
          "endsOn": {
            "type": "string",
            "example": "2026-12-18"
          },
          "days": {
            "type": "array",
            "uniqueItems": true,
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/Weekday"
            }
          },
          "startTime": {
            "type": "string",
            "description": "Dhaka wall-clock time, HH:MM (24-hour).",
            "example": "18:30"
          },
          "endTime": {
            "type": "string",
            "example": "20:30"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/BatchStatus"
              }
            ]
          },
          "seatsLeft": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "feeAmount": {
            "type": "number",
            "nullable": true,
            "description": "Whole taka; overrides the course price for this run. Null means the course price.",
            "minimum": 0
          }
        }
      },
...
    }
  }
}